(v0.3.0): switch from lazy to vim.pack
-Moved the plugin directory to the root directory
-Renamed the plugin directory to plugin.slink, so that
it's linked when installing the config
-Plugin updates
