--- lua.slink/plugin/blink.lua
@@ -1,18 +1,0 @@
-return {
-  {
-    'saghen/blink.cmp',
-    version = '*',
-    enabled = true,
-    opts = {
-      sources = {
-        default = { "lsp", "buffer" },
-      },
-      keymap = {
-        preset = 'super-tab'
-      },
-      fuzzy = {
-        implementation = 'prefer_rust'
-      }
-    }
-  }
-}
