--- lua.slink/plugin/colorscheme.lua
@@ -1,17 +1,0 @@
-return {
-  {
-    "rebelot/kanagawa.nvim",
-    lazy = false,
-    priority = 1000,
-  },
-  {
-    "vague2k/vague.nvim",
-    config = function()
-      -- NOTE: you do not need to call setup if you don't want to.
-      require("vague").setup({
-        -- optional configuration here
-      })
-      vim.cmd.colorscheme('vague')
-    end
-  },
-}
