--- lua.slink/plugin/lualine.lua
@@ -1,18 +1,0 @@
-return {
-  {
-    'nvim-lualine/lualine.nvim',
-    dependencies = { 'nvim-tree/nvim-web-devicons' },
-    config = function()
-      require('lualine').setup {
-        sections = {
-          lualine_a = { 'mode' },
-          lualine_b = { 'branch' },
-          lualine_c = { 'filename' },
-          lualine_x = { 'searchcount' },
-          lualine_y = { 'filetype' },
-          lualine_z = { 'location' }
-        },
-      }
-    end
-  }
-}
