+++ plugin.slink/blink.lua
@@ -1,0 +1,20 @@
+vim.pack.add({
+  "https://github.com/saghen/blink.lib",
+  "https://github.com/saghen/blink.cmp"
+})
+
+local cmp = require('blink.cmp')
+
+cmp.build():pwait()
+
+cmp.setup {
+  sources = {
+    default = { "lsp", "buffer" },
+  },
+  keymap = {
+    preset = 'super-tab'
+  },
+  fuzzy = {
+    implementation = 'prefer_rust'
+  }
+}
