+++ plugin.slink/treesitter.lua
@@ -1,0 +1,41 @@
+vim.pack.add({
+  "https://github.com/nvim-treesitter/nvim-treesitter"
+})
+
+local config = require("nvim-treesitter")
+
+config.setup {}
+config.install({
+  "c",
+  "cpp",
+  "lua",
+  "vim",
+  "tsx",
+  "json",
+  "go",
+  "rust",
+  "javascript",
+  "typescript",
+  "html",
+  "python",
+  "bash",
+  "vimdoc",
+  "java",
+  "query",
+  "markdown"
+})
+
+vim.api.nvim_create_autocmd('PackChanged', {
+  callback = function(ev)
+    local name, kind = ev.data.spec.name, ev.data.kind
+    if name == 'nvim-treesitter' and kind == 'update' then
+      if not ev.data.active then vim.cmd.packadd('nvim-treesitter') end
+      vim.cmd('TSUpdate')
+    end
+  end
+})
+
+-- vim.api.nvim_create_autocmd('FileType', {
+--   pattern = { '<filetype>' },
+--   callback = function() vim.treesitter.start() end,
+-- })
